Removed BaseMatrix.atIndex, .edIndex, .fnIndex
Replaced these with .at, .ed, .fn. Made relevant changes to other packages.

Internal to @envisim/matrix, replaced getters with actual values where so appropriate, such as this.nrow => this._nrow, as we should be able to use the underlying values directly.
This change, however, makes it more important to ensure that no accidental assignments are made to these (protected) properties.