Don't return multiple lines on intersectLineAreaFeatures (geojson-utils) (#17)
The current implementation of `intersectLineAreaFeatures` returns
multiple lines if the area-features are overlapping. This pull request
changes that, so that every line is checked against all area-features
for overlap, disregarding if areas themselves are overlapping.

In order to accomplish this effectively, a new class Segment is
introduced, which stores the line-equation for a segment, i.e. the
parametric equation between two `GeoJSON.Position`s.

Additionally, the proposed implementation has an improved efficiency
compared to the current implementation.