(geojson-utils, sampling) Removing enums (#88)
Replacing enums with objects in order to reduce reliance on typescript
code generation. This change makes it more likely to be able to run the
typescript code directly by a compiler that uses type stripping, such as
Node.

Replacing the BBox enums with a small class also responsible for
handling other smaller BBox-related methods.